+++ HISTORY.md
@@ -2,6 +2,7 @@ History of features added to LiChess Tools in time (reversed order)
 
 25 Apr
   - coords on each square of the board
+  - sound options: no move sounds
 
 16 Apr
   - Local evaluation extra chart